COM-3298 - create traineeTable
