+++ src/core/components/courses/TraineeContainer.vue
@@ -11,28 +11,9 @@
         <div v-else class="q-mb-sm">{{ maxTrainees }} stagiaires max</div>
       </div>
       <q-card>
-        <ni-responsive-table :data="course.trainees" :columns="traineesColumns" v-model:pagination="traineesPagination"
-          :visible-columns="traineesVisibleColumns" :loading="tableLoading">
-          <template #body="{ props }">
-            <q-tr :props="props">
-              <q-td v-for="col in props.cols" :key="col.name" :data-label="col.label" :props="props" :class="col.name"
-                :style="col.style">
-                <template v-if="col.name === 'actions'">
-                  <div class="row no-wrap table-actions">
-                    <ni-button icon="edit" @click="openTraineeEditionModal(props.row)"
-                      :disable="!!course.archivedAt" />
-                    <ni-button icon="close" @click="validateTraineeDeletion(col.value)"
-                      :disable="!!course.archivedAt" />
-                  </div>
-                </template>
-                <template v-else-if="!isClientInterface && col.name === 'company'">
-                  <div class="company-tag">{{ col.value }}</div>
-                </template>
-                <template v-else>{{ col.value }}</template>
-              </q-td>
-            </q-tr>
-          </template>
-        </ni-responsive-table>
+        <ni-trainee-table v-if="isIntraCourse" :trainees="course.trainees" :can-edit="canEdit" @refresh="refresh" />
+        <ni-expanding-table v-else :data="course.companies">
+        </ni-expanding-table>
         <q-card-actions align="right" v-if="canEdit">
           <ni-button color="primary" icon="add" label="Ajouter une personne" :disable="loading"
             @click="openTraineeCreationModal" />
@@ -43,9 +24,6 @@
     <trainee-addition-modal v-model="traineeAdditionModal" v-model:new-trainee="newTrainee" @submit="addTrainee"
       :validations="traineeValidation.newTrainee" :loading="traineeModalLoading" @hide="resetTraineeAdditionForm"
       :trainees-options="traineesOptions" @open-learner-creation-modal="openLearnerCreationModal" />
-
-    <trainee-edition-modal v-model="traineeEditionModal" v-model:edited-trainee="editedTrainee" @submit="updateTrainee"
-      @hide="resetTraineeEditionForm" :loading="traineeModalLoading" :validations="traineeValidation.editedTrainee" />
 
     <learner-creation-modal v-model="learnerCreationModal" v-model:new-user="newLearner" display-company
       @hide="resetLearnerCreationModal" :first-step="firstStep" @next-step="nextStepLearnerCreationModal"
@@ -58,10 +36,8 @@
 <script>
 import { computed, ref, toRefs } from 'vue';
 import { useStore } from 'vuex';
-import { useQuasar } from 'quasar';
 import get from 'lodash/get';
-import pick from 'lodash/pick';
-import omit from 'lodash/omit';
+import groupBy from 'lodash/groupBy';
 import Users from '@api/Users';
 import Courses from '@api/Courses';
 import {
@@ -73,17 +49,15 @@ import {
   AUXILIARY_WITHOUT_COMPANY,
 } from '@data/constants';
 import {
-  formatPhone,
-  formatPhoneForPayload,
   formatIdentity,
   formatAndSortOptions,
 } from '@helpers/utils';
 import Button from '@components/Button';
 import Input from '@components/form/Input';
-import ResponsiveTable from '@components/table/ResponsiveTable';
-import TraineeEditionModal from '@components/courses/TraineeEditionModal';
 import TraineeAdditionModal from '@components/courses/TraineeAdditionModal';
 import LearnerCreationModal from '@components/courses/LearnerCreationModal';
+import TraineeTable from '@components/courses/TraineeTable';
+import ExpandingTable from '@components/table/ExpandingTable';
 import { NotifyNegative, NotifyWarning, NotifyPositive } from '@components/popup/notify';
 import { useLearners } from '@composables/learners';
 import { useCourses } from '@composables/courses';
@@ -100,51 +74,17 @@ export default {
   components: {
     'ni-button': Button,
     'ni-input': Input,
-    'ni-responsive-table': ResponsiveTable,
-    'trainee-edition-modal': TraineeEditionModal,
+    'ni-trainee-table': TraineeTable,
     'trainee-addition-modal': TraineeAdditionModal,
     'learner-creation-modal': LearnerCreationModal,
+    'ni-expanding-table': ExpandingTable,
   },
   emits: ['refresh', 'update', 'update:maxTrainees'],
   setup (props, { emit }) {
     const { canEdit, validations, potentialTrainees } = toRefs(props);
 
     const $store = useStore();
-    const $q = useQuasar();
-    const traineesColumns = ref([
-      {
-        name: 'company',
-        label: 'Structure',
-        align: 'left',
-        field: row => get(row, 'company.name') || '',
-        classes: 'text-capitalize company-tag-content',
-      },
-      {
-        name: 'firstname',
-        label: 'Prénom',
-        align: 'left',
-        field: row => get(row, 'identity.firstname') || '',
-        classes: 'text-capitalize',
-      },
-      {
-        name: 'lastname',
-        label: 'Nom',
-        align: 'left',
-        field: row => get(row, 'identity.lastname') || '',
-        classes: 'text-capitalize',
-      },
-      { name: 'email', label: 'Email', align: 'left', field: row => get(row, 'local.email') || '' },
-      {
-        name: 'phone',
-        label: 'Téléphone',
-        align: 'left',
-        field: row => get(row, 'contact.phone') || '',
-        format: formatPhone,
-      },
-      { name: 'actions', label: '', align: 'left', field: '_id' },
-    ]);
-    const traineesPagination = ref({ rowsPerPage: 0, sortBy: 'lastname' });
-    const traineeEditionModal = ref(false);
+
     const traineeModalLoading = ref(false);
 
     const company = computed(() => $store.getters['main/getCompany']);
@@ -162,13 +102,6 @@ export default {
     const tableTitle = computed(() => (canEdit.value || isTrainer.value
       ? `Stagiaires (${traineesNumber.value})`
       : `Stagiaires de votre structure (${traineesNumber.value})`));
-
-    const traineesVisibleColumns = computed(() => {
-      const visibleColumns = ['firstname', 'lastname', 'email', 'phone'];
-      if (canEdit.value) visibleColumns.push('actions');
-
-      return isIntraCourse.value ? visibleColumns : ['company', ...visibleColumns];
-    });
 
     const traineesOptions = computed(() => potentialTrainees.value
       .map(pt => ({
@@ -193,19 +126,19 @@ export default {
 
     const companyOptions = computed(() => formatAndSortOptions(course.value.companies, 'name'));
 
+    const traineesGroupedByCompanies = computed(() => groupBy(course.value.trainees, t => t.company._id));
+
     const refresh = () => emit('refresh');
 
     const {
       newLearner,
       newTrainee,
       traineeAdditionModal,
-      editedTrainee,
       learnerCreationModal,
       learnerCreationModalLoading,
       firstStep,
       userAlreadyHasCompany,
       learnerAlreadyExists,
-      tableLoading,
       learnerValidation,
       traineeValidation,
       goToNextStep,
@@ -294,59 +227,6 @@ export default {
         traineeModalLoading.value = false;
       }
     };
-    const openTraineeEditionModal = async (trainee) => {
-      editedTrainee.value = {
-        ...editedTrainee.value,
-        ...pick(trainee, ['_id', 'identity.firstname', 'identity.lastname', 'local.email', 'contact.phone']),
-      };
-      traineeEditionModal.value = true;
-    };
-    const resetTraineeEditionForm = () => {
-      traineeValidation.value.editedTrainee.$reset();
-      editedTrainee.value = { identity: {}, local: {}, contact: {} };
-    };
-    const updateTrainee = async () => {
-      try {
-        traineeModalLoading.value = true;
-        traineeValidation.value.editedTrainee.$touch();
-        if (traineeValidation.value.editedTrainee.$error) return NotifyWarning('Champ(s) invalide(s)');
-        if (get(editedTrainee.value, 'contact.phone')) {
-          editedTrainee.value.contact.phone = formatPhoneForPayload(editedTrainee.value.contact.phone);
-        }
-
-        await Users.updateById(editedTrainee.value._id, omit(editedTrainee.value, ['_id', 'local']));
-        traineeEditionModal.value = false;
-
-        refresh();
-        NotifyPositive('Stagiaire modifié(e).');
-      } catch (e) {
-        console.error(e);
-        NotifyNegative('Erreur lors de la modification du/de la stagiaire.');
-      } finally {
-        traineeModalLoading.value = false;
-      }
-    };
-    const validateTraineeDeletion = (traineeId) => {
-      $q.dialog({
-        title: 'Confirmation',
-        message: 'Êtes-vous sûr(e) de vouloir retirer le/la stagiaire de la formation&nbsp;?',
-        html: true,
-        ok: true,
-        cancel: 'Annuler',
-      }).onOk(() => deleteTrainee(traineeId))
-        .onCancel(() => NotifyPositive('Suppression annulée.'));
-    };
-    const deleteTrainee = async (traineeId) => {
-      try {
-        await Courses.deleteTrainee(course.value._id, traineeId);
-
-        refresh();
-        NotifyPositive('Stagiaire supprimé(e).');
-      } catch (e) {
-        console.error(e);
-        NotifyNegative('Erreur lors de la suppression du/de la stagiaire.');
-      }
-    };
     const openTraineeCreationModal = () => {
       if (course.value.archivedAt) {
         return NotifyWarning('Vous ne pouvez pas ajouter de stagiaire à une formation archivée.');
@@ -367,16 +247,11 @@ export default {
       // Data
       newLearner,
       firstStep,
-      tableLoading,
       learnerCreationModalLoading,
       learnerCreationModal,
       learnerAlreadyExists,
       traineeAdditionModal,
       newTrainee,
-      editedTrainee,
-      traineesColumns,
-      traineesPagination,
-      traineeEditionModal,
       traineeModalLoading,
       companyOptions,
       // Validations
@@ -384,7 +259,6 @@ export default {
       traineeValidation,
       // Computed
       tableTitle,
-      traineesVisibleColumns,
       traineesOptions,
       disableCompany,
       course,
@@ -393,20 +267,18 @@ export default {
       isArchived,
       isRofOrAdmin,
       maxTraineesErrorMessage,
+      traineesGroupedByCompanies,
       // Methods
       nextStepLearnerCreationModal,
       submitLearnerCreationModal,
       resetLearnerCreationModal,
       resetTraineeAdditionForm,
       addTrainee,
-      openTraineeEditionModal,
-      resetTraineeEditionForm,
-      updateTrainee,
-      validateTraineeDeletion,
       openTraineeCreationModal,
       openLearnerCreationModal,
       updateMaxTrainees,
       inputTmpMaxTrainees,
+      refresh,
     };
   },
 };
